just use ubyte, we already depend on it via rocket

+++ src/orbit.rs
@@ -20,7 +20,6 @@ use libp2p::{
 };
 use rocket::tokio::task::JoinHandle;
 
-use byte_unit::Byte;
 use cached::proc_macro::cached;
 use std::{convert::TryFrom, error::Error as StdError, ops::Deref};
 
@@ -49,7 +48,7 @@ impl<T> Deref for AbortOnDrop<T> {
 
 #[derive(Clone, Debug, Default)]
 pub struct Limits {
-    storage: Option<Byte>,
+    storage: Option<usize>,
 }
 
 #[derive(Clone)]
